# Conflicts:
#	src/main/java/it/cnr/si/missioni/cmis/CMISOrdineMissioneService.java
#	src/main/java/it/cnr/si/missioni/service/RimborsoMissioneService.java
#	src/main/resources/config/application-cnr.yml
#	src/main/resources/it/cnr/missioni/sourceData/datiUo.json
#	src/main/resources/it/cnr/missioni/sourceData/usersSpecialForUo.json
#	src/main/resources/it/cnr/missioni/sourceData/usersSpecialForUoDev.json

+++ src/main/java/it/cnr/si/missioni/service/RimborsoMissioneDettagliService.java
@@ -184,7 +184,7 @@ public class RimborsoMissioneDettagliService {
 					livelloRich = 4;
 				}
 			} else {
-				if (tipoPasto.getLimiteMaxPasto().compareTo(new Double(50)) > 0){
+				if (tipoPasto.getLimiteMaxPasto().compareTo(new Double(30)) > 0){
 					livelloRich = 1;
 				} else {
 					livelloRich = 4;
